[AC-1124] Moved UpdatePaymentForm to origin place
